🚛 rename package to pruefi
